refactor(Paging): filterPaging function now in correct file
