+++ src/Ml.Cli.WebApp/ClientApp/src/Comparison/TableResult.js
@@ -3,36 +3,8 @@ import {Paging} from "@axa-fr/react-toolkit-table";
 import TableItem from './TableItem';
 import StatsTable from './StatsTable';
 import {formatJson} from "./FormatFilter";
-import {computeNumberPages} from "../Tables/Paging";
+import {computeNumberPages, filterPaging} from "../Tables/Paging";
 import './TableResult.scss';
-
-export const filterPaging = (items, numberItems, currentPage) => {
-    let pageIndex = 0;
-    if (currentPage > 1) {
-        pageIndex = currentPage - 1;
-    }
-
-    let controlledNumberItems = 0;
-    if (numberItems > 0) {
-        controlledNumberItems = numberItems;
-    }
-
-    const {length} = items;
-    let beginIndex = pageIndex * controlledNumberItems;
-    const lastIndex = length < beginIndex + controlledNumberItems ? length : beginIndex + controlledNumberItems;
-
-    let returnedCurrentPage = currentPage;
-
-    if (lastIndex < beginIndex) {
-        if (lastIndex - length >= 0) {
-            beginIndex = lastIndex - length;
-        } else {
-            beginIndex = 0;
-        }
-        returnedCurrentPage = -1;
-    }
-    return {items: items.slice(beginIndex, lastIndex), currentPage: returnedCurrentPage};
-};
 
 export const filterItems = (items, filterName) => {
     return items.filter(item => {
